backend/requirements.txt
Django==5.0.7
djangorestframework==3.15.2
psycopg2-binary==2.9.9
celery==5.4.0
redis==5.0.7
python-dotenv==1.0.1
PyGithub==2.4.0
requests==2.32.3

# Static analysis
bandit==1.7.9
flake8==7.1.1
semgrep==1.91.0
safety==3.2.4

# Testing
pytest==8.3.2
pytest-cov==5.0.0
hypothesis==6.112.2
junitparser==3.2.0

gunicorn

# Optional model clients (simple HTTP)
